latest

+++ Models/RequestModels/RegisterUser.cs
@@ -22,7 +22,7 @@ namespace Models.RequestModels
         public List<SelectItem> Roles { get; set; }
 
         [Required(ErrorMessage = "Please Choose Language")]
-        public long Language { get; set; }
+        public int Language { get; set; }
 
         public List<SelectItem> Languages { get; set; }
